Extract client to module
